Rumor

+++ Peerster/Peer.cc
@@ -8,7 +8,7 @@
 
 #include "Peer.hh"
 
-Peer::Peer(QHostAddress a, quint32 p) {
+Peer::Peer(QHostAddress a, quint16 p) {
     address = a;
     port = p;
 }
